faq: fix duplicates on tab switching
`onCreateView` will run more than once during the lifetime of the
fragment. This means we can't add the answers and questions to a list on
the object on each `onCreateView`.

Instead we initialize the questions and answers lists when declaring
them. We can then also use `listOf` instead of `mutableListOf`.