map: add hack to avoid redoing state on `onCreateView`
The map handles most state itself, through `savedInstanceState`. So
on onCreateView we are in danger of re-adding things the map already
have re-instantiated, like polygons or cameare position.

We can detect if this is the case through the existence of
savedInstanceState. If it's non-null in onCreateView we know the map
will handle most things and we can mostly defer to the map.