map: lookup weather/address in the viewmodel
We can express dependencies between livedata using transformations. This
is really what we want to do with `weather` and `address` as they depend
necessarily depend on `position`.

So when `position` is set the viewmodel will react and run the
approriate code to fetch the `address` and `weather` in a coroutine. The
only thing we need to do in HomeFragment is set the position
and then wait until the livadata in `weather` and `address` is changed.

The main immediate benefit is not delaying the popup when we come back
to the map. The weather/address information is already there so it can
be populated immediatedly.

In the future we can add a refresh timer in the viewmodel which then
refreshes the weather information to make it timely.