map: simplify setMarker and observers
`setMarker` no longer does anything that can't be in `onMapClick` so
move the code there.

We no longer need to observe `position` as that's done in the viewmodel.
And there's no need to add the observers in `onmapreadycallback` so we
do that in `onCreateView` (the documentation says this should be done in
`onCreate`, but we don't have access to the lifecycle object yet apparantly).