recyclerViews: preserve scroll positions on change
We currently recreate the layout and adapter each time the underlying
live data changes, this kills the scroll position.

Instead just notify the adapter that the data has changed.

Note: for this to work we always need to modify the underlying list with
eg. add/remove. If we pass an entirely new list to the livedata the
adapter isn't able to see anything.

Alternatively it might be possible to expose the adapters list and reset
only that.

fixes #59